Separate Spring Schduler test from SpringBootAuto test
